refactor: bump onyxcord and complete Discord application commands
Due to a recent update of onyxcord and a fusion of structures, all `bot.Client` calls have been replaced to simply `bot`, and the CLI has been implemented. The list of application commands that must be registered to Discord has also been updated to feature all of them, so we can properly use the `register` CLI command.