feat: add expiration time for games
Added expirations for the games so they don't last forever. The mechanism is fairly simple : the Redis key associated with every game is set to expire about a certain amount of time (renewed at each player participation). When it's the case it sends a message through a Pub/Sub channel, received by the bot so it can send a message to tell the game has expired.